create test case for controller and service

+++ src/main/java/com/washingmachine/service/WashingMachineService.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Service;
 import com.washingmachine.exception.WasinghMachineAlreadyExistException;
 import com.washingmachine.exception.WasinghMachineNotFoundException;
 import com.washingmachine.model.WashingMachine;
+import com.washingmachine.model.WashingMachineStatus;
 import com.washingmachine.repository.WashingMachineRepository;
 
 @Service
@@ -32,7 +33,7 @@ public class WashingMachineService {
 	public WashingMachine getWashingMachineById(Long id) {
 
 		Optional<WashingMachine> waMachine = washingMachineRepository.findById(id);
-		if (waMachine.isEmpty()) {
+		if (!waMachine.isPresent()) {
 			throw new WasinghMachineNotFoundException("This Washing Machine Not Found for this Id =" + id);
 		}
 		return waMachine.get();
@@ -49,11 +50,26 @@ public class WashingMachineService {
 	}
 	
 	public void delete(Long id) {
+		Optional<WashingMachine> waMachine = washingMachineRepository.findById(id);
+		if (!waMachine.isPresent()) {
+			throw new WasinghMachineNotFoundException("This Washing Machine Not Found for this Id =" + id);
+		}
 		washingMachineRepository.deleteById(id);
 	}
 	
 	public void update(WashingMachine washingMachine,Long washingMachineId) {
+		
 		WashingMachine washingMachine2 = washingMachineRepository.findById(washingMachineId).get();
+		
+		if(washingMachine.getStatus().equals(WashingMachineStatus.WASHING)) {
+			if(washingMachine2.getStatus().equals(WashingMachineStatus.STOP)) {
+				throw new WasinghMachineNotFoundException("You can't wash because your machine is not start");
+			}
+		}else if(washingMachine.getStatus().equals(WashingMachineStatus.DRYING)) {
+			if(washingMachine2.getStatus().equals(WashingMachineStatus.STOP)) {
+				throw new WasinghMachineNotFoundException("You can't wash because your machine is not start");
+			}
+		}
 		washingMachineRepository.save(washingMachine);
 	}
 }
